Remove usage of SharedBuffer
Last year didn't have much, but one thing it had was that `SharedBuffer` worked in Chrome. Now it doesn't, I guess for reasons related to Spectre. So instead of using SharedBuffer to control pause/resume, just use messages. Worker loop yields every so often to allow messages to be processed.

Also fix a typo in examples. Also, since 2020 is over, update the description to describe the zeitgeist when this was made (even though it's not as better as it should be, even now)